requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
djangorestframework==3.12.4
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
Pillow==8.2.0
pydub==0.25.1
python-dateutil==1.5
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
waitress==2.0.0
whitenoise==5.2.0
